backend/requirements.txt
fastapi
uvicorn[standard]
groq
google-generativeai
